+++ tests/LinkSharerTest.php
@@ -22,7 +22,6 @@ class LinkSharerTest extends TestCase
 
         self::assertSame('https://twitter.com/intent/tweet?url=https%3A%2F%2Fexample.com%2Fpath%2Ffile', $linkSharer->getTwitterSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.facebook.com/sharer/sharer.php?u=https%3A%2F%2Fexample.com%2Fpath%2Ffile', $linkSharer->getFacebookSharer()->getShareUrl()->__toString());
-        self::assertSame('https://plus.google.com/share?url=https%3A%2F%2Fexample.com%2Fpath%2Ffile', $linkSharer->getGooglePlusSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.linkedin.com/shareArticle?mini=true&url=https%3A%2F%2Fexample.com%2Fpath%2Ffile', $linkSharer->getLinkedInSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.reddit.com/submit?url=https%3A%2F%2Fexample.com%2Fpath%2Ffile', $linkSharer->getRedditSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.tumblr.com/widgets/share/tool?canonicalUrl=https%3A%2F%2Fexample.com%2Fpath%2Ffile', $linkSharer->getTumblrSharer()->getShareUrl()->__toString());
@@ -37,7 +36,6 @@ class LinkSharerTest extends TestCase
 
         self::assertSame('https://twitter.com/intent/tweet?url=https%3A%2F%2Fexample.com%2Fpath%2Ffile&text=I%20am%20sharing%20this.', $linkSharer->getTwitterSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.facebook.com/sharer/sharer.php?u=https%3A%2F%2Fexample.com%2Fpath%2Ffile', $linkSharer->getFacebookSharer()->getShareUrl()->__toString());
-        self::assertSame('https://plus.google.com/share?url=https%3A%2F%2Fexample.com%2Fpath%2Ffile', $linkSharer->getGooglePlusSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.linkedin.com/shareArticle?mini=true&url=https%3A%2F%2Fexample.com%2Fpath%2Ffile&title=I%20am%20sharing%20this.', $linkSharer->getLinkedInSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.reddit.com/submit?url=https%3A%2F%2Fexample.com%2Fpath%2Ffile&title=I%20am%20sharing%20this.', $linkSharer->getRedditSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.tumblr.com/widgets/share/tool?canonicalUrl=https%3A%2F%2Fexample.com%2Fpath%2Ffile&title=I%20am%20sharing%20this.', $linkSharer->getTumblrSharer()->getShareUrl()->__toString());
@@ -52,7 +50,6 @@ class LinkSharerTest extends TestCase
 
         self::assertSame('https://twitter.com/intent/tweet?url=https%3A%2F%2Fexample.com%2Fpath%2Ffile&text=I%20am%20sharing%20this.&hashtags=sharing%2Cthis', $linkSharer->getTwitterSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.facebook.com/sharer/sharer.php?u=https%3A%2F%2Fexample.com%2Fpath%2Ffile&hashtag=%23sharing', $linkSharer->getFacebookSharer()->getShareUrl()->__toString());
-        self::assertSame('https://plus.google.com/share?url=https%3A%2F%2Fexample.com%2Fpath%2Ffile', $linkSharer->getGooglePlusSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.linkedin.com/shareArticle?mini=true&url=https%3A%2F%2Fexample.com%2Fpath%2Ffile&title=I%20am%20sharing%20this.', $linkSharer->getLinkedInSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.reddit.com/submit?url=https%3A%2F%2Fexample.com%2Fpath%2Ffile&title=I%20am%20sharing%20this.', $linkSharer->getRedditSharer()->getShareUrl()->__toString());
         self::assertSame('https://www.tumblr.com/widgets/share/tool?canonicalUrl=https%3A%2F%2Fexample.com%2Fpath%2Ffile&title=I%20am%20sharing%20this.&tags=sharing%2Cthis', $linkSharer->getTumblrSharer()->getShareUrl()->__toString());